+++ src/libchips/Makefile
@@ -7,13 +7,13 @@ all: libchips.so.1.1 libchips_ce.so.1.1
 %.o: %.c
 	$(CC) $(GDALCFLAGS) $(CFLAGS) -fPIC $< -c -o $@
 
-chips_ce.o: chips.c
+chips_ce.o: chips.c globals.o
 	$(CC) $(GDALCFLAGS) $(CFLAGS) -DCHAMPION_EDITION=1 -fPIC $< -c -o $@
 
 %.o: %.cpp
 	$(CXX) $(GDALCFLAGS) $(CXXFLAGS) $(CFLAGS) -fPIC $< -c -o $@
 
-libchips.so.1.1: chips.o
+libchips.so.1.1: chips.o globals.o
 	$(CC) $(CFLAGS) $^ $(LDFLAGS) -shared -o $@
 	strip $@
 
@@ -21,8 +21,8 @@ libchips_ce.so.1.1: chips_ce.o
 	$(CC) $(CFLAGS) $^ $(LDFLAGS) -shared -o $@
 	strip $@
 
-main: main.c chips.c chips.h
-	$(CC) $(GDALCFLAGS) $(CFLAGS) -I . main.c chips.c $(shell pkg-config gdal --libs) -lpthread -o $@
+main: main.c chips.c globals.c
+	$(CC) $(GDALCFLAGS) $(CFLAGS) -I . main.c chips.c globals.c $(shell pkg-config gdal --libs) -lpthread -o $@
 
 clean:
 	rm -f *.o
