+++ src/libchips/chips.h
@@ -2,7 +2,7 @@
  * The MIT License (MIT)
  * =====================
  *
- * Copyright © 2019 Azavea
+ * Copyright © 2019-2020 Azavea
  *
  * Permission is hereby granted, free of charge, to any person
  * obtaining a copy of this software and associated documentation
